requirements.txt
openai
streamlit
streamlit-extras
streamlit_cookies_controller
python-dotenv